add alternative-rewrite-activation, allow sql input to commandline

+++ include/common.h
@@ -1,7 +1,7 @@
 #ifndef COMMON_H
 #define COMMON_H
 
-typedef int boolean;
+typedef short boolean;
 #define TRUE 1
 #define FALSE 0
